+++ NEWS.md
@@ -1,3 +1,17 @@
+## version 2.3.0
+
+---
+
+### Updates
+
+- length of boundaries is calculated as distance from farthest points through edges of a subgraph of adjacent boundary elements, instead of straight line between points
+- Gaussian neutral model uses autocorrelation range based on variogram, rather than LISA clusters
+- categorical_boundary function deprecated, and functionality merged with define_boundary function
+- renamed statistical test functions to improve clarity
+- plot_boundary now optionally returns a SpatRaster object with boundary elements and boundary element overlaps
+- removed dependencies on sf and pdqr
+- added dependency on gstat
+
 ## version 2.2.0
 
 ---